requirements.txt
tensorflow
pandas
numpy
SQLAlchemy
mysqlclient
fastapi
uvicorn[standard]
python-dotenv
pydantic
fastapi-redis-cache
gdown
haversine